feat: support dlx/npx again
