--- bin/scripts/invoke.js
@@ -1,12 +1,0 @@
-const spawn = require('cross-spawn');
-
-const invoke = () => spawn.sync(
-    'mrm',
-    [
-        'main', '--preset', '@solid-soda/scripts',
-    ],
-    { stdio: 'inherit' },
-)
-
-
-module.exports = invoke
