+++ bin/solid-scripts.js
@@ -2,25 +2,45 @@
 
 const commandLineArgs = require('command-line-args')
 
-const invoke = require('./scripts/invoke')
-const update = require('./scripts/update')
-
 const definitions = [
     { name: 'command', defaultOption: true }
 ]
 
-const options = commandLineArgs(definitions, { stopAtFirstUnknown: true });
 
+const update = ({ withSelf } = {}) => {
+    const packages = ['mrm']
+
+    if (withSelf) {
+        packages.push('@solid-soda/scripts')
+    }
+
+    spawn.sync(
+        'npm',
+        [
+            'install',
+            '-g',
+            ...packages,
+        ],
+        { stdio: 'inherit' },
+    )
+
+    spawn.sync(
+        'mrm',
+        [
+            'main', '--preset', '@solid-soda/scripts',
+        ],
+        { stdio: 'inherit' },
+    )
+}
+
+const options = commandLineArgs(definitions, { stopAtFirstUnknown: true });
 switch (options.command) {
     case 'init':
     case 'update':
         update({ withSelf: true })
         break
-    case 'x':
-        update({ withSelf: false })
-        break
     default:
-        invoke()
+        update({ withSelf: false })
         break;
 }
